chore: added fixes from comments
