+++ packages/onboard-react/src/components/HardwareWalletSteps/DerivationPathForm.module.css
@@ -1,6 +1,9 @@
 .container {
-  @apply flex flex-col justify-between px-6;
+  @apply flex flex-col justify-between;
   min-height: 250px;
+  @screen md {
+    @apply px-6;
+  }
 }
 
 .infoText {
