+++ packages/onboard-react/src/components/HardwareWallets/HardwareWallets.module.css
@@ -1,0 +1,3 @@
+.closeButton {
+  @apply absolute top-16 left-6;
+}
