+++ packages/onboard-react/src/components/HardwareWallets/HardwareWallets.tsx
@@ -1,4 +1,4 @@
-import { FC, useCallback, useState } from "react";
+import { FC, useCallback, useMemo, useState } from "react";
 import { closeAccountSelect, selectAccounts$ } from "@sovryn/onboard-hw-common";
 import { useSubscription } from "../../hooks/useSubscription";
 import { FilterType, WalletList } from "../WalletList/WalletList";
@@ -8,10 +8,13 @@ import {
   HardwareWalletSteps,
 } from "../HardwareWalletSteps/HardwareWalletSteps";
 import { ButtonBack } from "../ButtonBack/ButtonBack";
+import styles from "./HardwareWallets.module.css";
+import { useIsMobile } from "../../hooks/useIsMobile";
 
 export const HardwareWallets: FC = () => {
   const { inProgress } = useSubscription(selectAccounts$);
   const [step, setStep] = useState(HardwareWalletStep.derivationPathForm);
+  const { isMobile } = useIsMobile();
 
   const handleStepBack = useCallback(() => {
     if (step === HardwareWalletStep.derivationPathForm) {
@@ -21,13 +24,28 @@ export const HardwareWallets: FC = () => {
     }
   }, [step]);
 
+  const renderButtonBack = useMemo(() => {
+    const buttonText =
+      step === HardwareWalletStep.addressList
+        ? "Back to derivation path"
+        : "Back to wallet menu";
+
+    return (
+      <ButtonBack
+        label={isMobile ? buttonText : ""}
+        className={!isMobile && styles.closeButton}
+        onClick={handleStepBack}
+      />
+    );
+  }, [step, isMobile, handleStepBack]);
+
   return (
     <>
       {!inProgress ? (
         <WalletList filter={FilterType.hardware} />
       ) : (
         <>
-          <ButtonBack label="Back to wallet menu" onClick={handleStepBack} />
+          {renderButtonBack}
           <div>
             <HardwareWalletSteps step={step} onStepChanged={setStep} />
           </div>
